+++ runtests.py
@@ -72,4 +72,5 @@ def start():
         exit(0)
 
 if __name__ == '__main__':
+    print(sys.version)
     start()